Converted the project into typescript
