+++ package.json
@@ -5,11 +5,17 @@
   "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start-server": "node app.js"
+    "start-server": "node dist/app.js",
+    "compile": "rm -rf dist && npx tsc",
+    "dev": "npm run compile && npm run start-server"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "express": "^4.18.2"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.17",
+    "typescript": "^5.1.6"
   }
 }
