+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+      "module": "commonjs",
+      "esModuleInterop": true,
+      "target": "es6",
+      "sourceMap": true,
+      "outDir": "dist"
+  }
+}
